+++ src/main/java/api/m2/movements/enums/MovementType.java
@@ -3,5 +3,6 @@ package api.m2.movements.enums;
 public enum MovementType {
     DEBITO,
     CREDITO,
-    INGRESO
+    INGRESO,
+    REINTEGRO
 }